+++ lib/llm_composer/provider_response/bedrock.ex
@@ -1,0 +1,7 @@
+defmodule LlmComposer.ProviderResponse.Bedrock do
+  @moduledoc false
+
+  use LlmComposer.ProviderResponse.Struct,
+    parser: LlmComposer.ProviderResponse.Parser.Bedrock,
+    provider: :bedrock
+end