+++ lib/llm_composer/provider_response/google.ex
@@ -1,0 +1,7 @@
+defmodule LlmComposer.ProviderResponse.Google do
+  @moduledoc false
+
+  use LlmComposer.ProviderResponse.Struct,
+    parser: LlmComposer.ProviderResponse.Parser.Google,
+    provider: :google
+end