+++ lib/llm_composer/provider_response/open_router.ex
@@ -1,0 +1,7 @@
+defmodule LlmComposer.ProviderResponse.OpenRouter do
+  @moduledoc false
+
+  use LlmComposer.ProviderResponse.Struct,
+    parser: LlmComposer.ProviderResponse.Parser.OpenAI,
+    provider: :open_router
+end