+++ lib/llm_composer/provider_stream_chunk/open_ai_responses.ex
@@ -1,0 +1,7 @@
+defmodule LlmComposer.ProviderStreamChunk.OpenAIResponses do
+  @moduledoc false
+
+  use LlmComposer.ProviderStreamChunk.Struct,
+    parser: LlmComposer.ProviderStreamChunk.Parser.OpenAIResponses,
+    provider: :open_ai_responses
+end